Populate compiler path setting in template
Currently the project template includes a fixed path to a named user
directory which won't be valid on another machine. This has the effect
that it will fall back to the system GCC, which doesn't know about Amiga
libraries etc.

I've changed this to a placeholder which gets replaced with the path of
the bundled GCC binary.