Absolute position decoration workaround
Prevents misalignment / jumping while editing due to the delay adding
decorations. The idea is that *every* line always has an 'empty'
decoration to add horizontal space. We can add this immediately to new
lines on edit. The actual cycle count decoration is added in addition to
this and absolutely positioned on top, so doesn't affect layout.